requirements.txt
pycryptodomex>=3.15.0
tabulate>=0.8.9
matplotlib>=3.5.0
